refactor: step3 요구사항 형식에 맞게 수정
