+++ tomcat/src/main/java/org/apache/catalina/Controller.java
@@ -1,0 +1,10 @@
+package org.apache.catalina;
+
+import java.io.IOException;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
+
+public interface Controller {
+
+    void service(HttpRequest request, HttpResponse response) throws IOException;
+}
